+++ src/app/modules/global-admin/global-admin-organizations/organizations-dialogs/delete-organization-dialog/removal-conflict-table/inner-conflict-table/inner-conflict-table.component.ts
