+++ src/app/shared/models/global-admin/organization-removal-conflicts.model.ts
@@ -1,4 +1,4 @@
-import { RemovalConflict } from 'src/app/modules/global-admin/global-admin-organizations/delete-organization-dialog/removal-conflict-table/removal-conflict-table.component';
+import { RemovalConflict } from 'src/app/modules/global-admin/global-admin-organizations/organizations-dialogs/delete-organization-dialog/removal-conflict-table/removal-conflict-table.component';
 
 export interface OrganizationRemovalConflicts {
   contractsInOtherOrganizationsWhereOrgIsSupplier: RemovalConflict[];
